requirements.txt
mistralai==1.5.0
pinecone==5.0.1
psycopg2_binary==2.9.9
python-dotenv==1.0.1
python-telegram-bot==21.10
Requests==2.32.3
selenium==4.28.1
